댓글쓰기 fetch 요청 및 커스텀 익셉션 만들고 등록하고 삭제하고 cascade와 fk null변경

+++ src/main/java/shop/mtcoding/blogv2/reply/ReplyService.java
@@ -1,40 +1,73 @@
 package shop.mtcoding.blogv2.reply;
-
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import shop.mtcoding.blogv2._core.error.ex.MyApiException;
+import shop.mtcoding.blogv2._core.error.ex.MyException;
 import shop.mtcoding.blogv2.board.Board;
+import shop.mtcoding.blogv2.board.BoardRepository;
+import shop.mtcoding.blogv2.reply.ReplyRequest.SaveDTO;
 import shop.mtcoding.blogv2.user.User;
-
 @Service
 public class ReplyService {
-
     @Autowired
     private ReplyRepository replyRepository;
-
+    @Autowired
+    private BoardRepository boardRepository;
     @Transactional
-    public void 삭제하기(Integer id) {
-        try {
-            replyRepository.deleteById(id);
-        } catch (Exception e) {
-            throw new RuntimeException(id + "를 찾을 수 없어요");
-        }
-    }
-
-    @Transactional
-    public void 댓글쓰기(ReplyRequest.SaveDTO saveDTO, int sessionUserId) {
-
+    public void 댓글쓰기(SaveDTO saveDTO, Integer sessionId) {
+        // insert into reply_tb(comment, board_id, user_id) values(?,?,?)
         Board board = Board.builder().id(saveDTO.getBoardId()).build();
-        Reply reply = Reply.builder()
-                .comment(saveDTO.getComment())
-                .board(board)
-                // .board(saveDTO.getBoardId()) // DTO를 오브젝트로.
-                // .board(Board.builder().id(saveDTO.getBoardId()).build()) // 바로 빌더. 생성시점따라 오류가능성있음
-                .user(User.builder().id(sessionUserId).build())
-                .build();
+        User user = User.builder().id(sessionId).build();
 
-        replyRepository.save(reply);
-    }
+        Reply reply = Reply.builder()
+        .comment(saveDTO.getComment())
+        .board(board)
+        .user(user)
+        .build();
+replyRepository.save(reply); // entity : Reply 객체
+}
+
+@Transactional
+public void 댓글삭제(Integer id, Integer sessionUserId) {
+Optional<Reply> replyOP = replyRepository.findById(id);
+
+if (replyOP.isEmpty()) {
+    throw new MyApiException("삭제할 댓글이 없습니다");
+}
+
+Reply reply = replyOP.get();
+if (reply.getUser().getId() != sessionUserId) {
+    throw new MyApiException("해당 댓글을 삭제할 권한이 없습니다");
+}
+
+replyRepository.deleteById(id);
+}
 
 }
+
+// @Transactional
+// public void 삭제하기(Integer id) {
+// try {
+// replyRepository.deleteById(id);
+// } catch (Exception e) {
+// throw new RuntimeException(id + "를 찾을 수 없어요");
+// }
+// }
+// @Transactional
+// public void 댓글쓰기(ReplyRequest.SaveDTO saveDTO, int sessionUserId) {
+
+// Board board = Board.builder().id(saveDTO.getBoardId()).build();
+// Reply reply = Reply.builder()
+// .comment(saveDTO.getComment())
+// .board(board)
+// // .board(saveDTO.getBoardId()) // DTO를 오브젝트로.
+// // .board(Board.builder().id(saveDTO.getBoardId()).build()) // 바로 빌더. 생성시점따라
+// 오류가능성있음
+// .user(User.builder().id(sessionUserId).build())
+// .build();
+
+// replyRepository.save(reply);
+// }
